Fixed camera buffering issue
Michael Kelly reported that the camera was bufferin up all of the images
and then causing the event to fire very quickly. As a result, the GUI
was crashing. This fix first empties the camera buffer before it
starts notifying listeners.

The fix is to keep pulling a frame until the time to take the frame
is less than 10 milliseconds.

In the future, we may want to update OPENCV to address this issue