Add Command Framework
Updating commands has been moved to its own method to allow us to run Commands in teleop without repeating code (last year this was done in the DriveTrain class). The logic has been slightly tweaked so the current Command does not have to be initialized separately at the start of auto.